requirements.txt
numpy==1.22.3
pandas==1.4.2
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
speedtest-cli==2.1.3
